Fixed Weather API parsing

+++ src/main/java/me/firestone82/solaxautomation/service/meteosource/model/type/Cloud.java
@@ -4,5 +4,5 @@ import lombok.Data;
 
 @Data
 public class Cloud {
-    private final int total;
+    private final double total;
 }
